requirements.txt
streamlit
Pillow
torch
torchvision
python-dotenv
langchain-groq
requests
matplotlib
tensorflow
